feat: 게시물 삭제 상태 변경 기능 추가 fix: 스크롤 기능의 응답데이터 내 TotalElements 정보 수정 (Post Profile, find Comments, find Dms, find DmGroups, find RecentSearches) refactor: Request, Response Inner Entity로 정리 (Post, PostImage, PostLike)

+++ src/main/java/com/project/danstagram/domain/search/dto/SearchResponse.java
@@ -16,8 +16,7 @@ public class SearchResponse {
 
     /* findRecentSearches */
     @Builder
-    public record RecentSearchList(int totalElements, Long nextCursor, List<RecentSearches> recentSearches) {
-
+    public record RecentSearchList(Long totalElements, Long nextCursor, List<RecentSearches> recentSearches) {
     }
 
     public record RecentSearches(Long searchIdx, String searchDate, String searchType, String searchKeyword) {
